pasar de carrito a facturas, aplicacion de permisos para inventario y pedido

+++ app/Http/Controllers/mercancia/MprArticuloController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\mercancia;
 
+use App\Http\Controllers\administracion\musUsuarioController;
 use App\Http\Controllers\Controller;
 use App\Http\Controllers\MprFechaUpdateContoller;
 use App\Models\administracion\MadComentario;
@@ -10,9 +11,7 @@ use App\Models\mercancia\MprArticulo;
 use App\Models\mercancia\MprImagen;
 use App\Models\mercancia\MprMarca;
 use App\Models\mercancia\MprtpArticulo;
-use App\Models\venta\MveCarrito;
-use App\Models\venta\MveDetCarrito;
-
+use Illuminate\Support\Facades\Auth;
 
 class MprArticuloController extends Controller
 {
@@ -24,25 +23,36 @@ class MprArticuloController extends Controller
     public function index()
     {
         // $articulos = MprArticulo::all();
-        $articulos = MprArticulo::where('estado',1)->get();
+        $articulos = MprArticulo::where('estado', 1)->get();
 
         $tipo_articulo = MprtpArticulo::all()
-        ->where('estado',1);
+            ->where('estado', 1);
 
-        foreach($articulos as $articulo){
-            $imagenes[$articulo->id] = MprImagen::select('id','ruta','articulo')
-            ->where('articulo',$articulo->id)
-            ->where('estado',1)
-            ->first();
+        foreach ($articulos as $articulo) {
+            $imagenes[$articulo->id] = MprImagen::select('id', 'ruta', 'articulo')
+                ->where('articulo', $articulo->id)
+                ->where('estado', 1)
+                ->first();
         }
-        return view('mercancia.mprarticulos.index',compact('articulos','tipo_articulo','imagenes'));
+        return view('mercancia.mprarticulos.index', compact('articulos', 'tipo_articulo', 'imagenes'));
     }
 
     public function adminArticulos()
     {
-        // $articulos = MprArticulo::all();
-        $articulos = MprArticulo::where('estado',1)->get();
-        return view('mercancia.mprarticulos.AdminArticulos',compact('articulos'));
+        if (Auth::guard('usuarios')->user()) {
+            $ctru = new musUsuarioController();
+            if ($ctru->getPermisoInv()->leer == 1) {
+                $articulos = MprArticulo::where('estado', 1)->get();
+            
+                for ($i=0; $i < count($articulos); $i++) { 
+                   if( strlen($articulos[$i]->descripcion)>30){
+                    $articulos[$i]->descripcion = substr( $articulos[$i]->descripcion , 0, 20) . "...";
+                   }
+                }
+                return view('mercancia.mprarticulos.AdminArticulos', compact('articulos'));
+            }
+        }
+        return redirect()->route('homeAdmin');
     }
 
     /**
@@ -52,10 +62,20 @@ class MprArticuloController extends Controller
      */
     public function create()
     {
-        $tmarticulos =  MprMarca::all();
-        $tparticulos = MprtpArticulo::all();
-        return view ('mercancia.mprarticulos.crear',compact('tmarticulos','tparticulos'));
+        if (Auth::guard('usuarios')->user()) {
+            $ctru = new musUsuarioController();
+            if ($ctru->getPermisoInv()->crear == 1) {
+                $tmarticulos =  MprMarca::all();
+                $tparticulos = MprtpArticulo::all();
+                return view('mercancia.mprarticulos.crear', compact('tmarticulos', 'tparticulos'));
+            }
+        }
+        return redirect()->route('homeAdmin');
     }
+
+
+
+
 
     /**
      * Store a newly created resource in storage.
@@ -65,25 +85,30 @@ class MprArticuloController extends Controller
      */
     public function store(Request $request)
     {
-        $nuevo = MprArticulo::create ([
-            'nombre' => $request -> nombre,
-            'descripcion' => $request -> descripcion,
-            'precio' => $request -> precio,
-            'existencia' => $request -> existencia,
-            'marca' => $request -> marca,
-            'vista'=>"0",
-            'tipoarticulo' => $request -> tipoarticulo,
-            'estado' => "1"
-        ]);
+        if (Auth::guard('usuarios')->user()) {
+            $ctru = new musUsuarioController();
+            if ($ctru->getPermisoInv()->crear == 1) {
+                $nuevo = MprArticulo::create([
+                    'nombre' => $request->nombre,
+                    'descripcion' => $request->descripcion,
+                    'precio' => $request->precio,
+                    'existencia' => $request->existencia,
+                    'marca' => $request->marca,
+                    'vista' => "0",
+                    'tipoarticulo' => $request->tipoarticulo,
+                    'estado' => "1"
+                ]);
 
-        $comboImg = MprImagen::where('estado', 2)->get();
-        for ($i = 0; $i < count($comboImg); $i++) {
-            $comboImg[$i]->articulo = $nuevo->id;
-            $comboImg[$i]->estado = 1;
-            $comboImg[$i]->save();
+                $comboImg = MprImagen::where('estado', 2)->get();
+                for ($i = 0; $i < count($comboImg); $i++) {
+                    $comboImg[$i]->articulo = $nuevo->id;
+                    $comboImg[$i]->estado = 1;
+                    $comboImg[$i]->save();
+                }
+                return redirect()->route('adminArticulos');
+            }
         }
-
-        return redirect()->route('adminArticulos');
+        return redirect()->route('homeAdmin');
     }
 
     /**
@@ -94,45 +119,56 @@ class MprArticuloController extends Controller
      */
     public function show($id)
     {
+
         $articulo = MprArticulo::findOrFail($id);
+        if (Auth::user()) {
+            $articulo->vistas = $articulo['vistas'] + 1;
+        }
+        $articulo->update();
         $marca = MprMarca::findOrFail($articulo->marca);
         $tparticulos = MprtpArticulo::findOrFail($articulo->tipoarticulo);
+        $relacionados = MprArticulo::select('id', 'nombre', 'precio')
+            ->where('tipoarticulo', $articulo->tipoarticulo)
+            ->where('id', '<>', $id)
+            ->where('estado', 1)->get();
 
-        $relacionados = MprArticulo::select('id','nombre','precio')
-        ->where('tipoarticulo',$articulo->tipoarticulo)
-        ->where('id', '<>',$id)
-        ->where('estado',1)->get();
-       
-        foreach($relacionados as $relArticulo){
-            $relImagenes[$relArticulo->id] = MprImagen::select('id','ruta','articulo')
-            ->where('articulo',$relArticulo->id)
-            ->where('estado',1)
-            ->first();
+        foreach ($relacionados as $relArticulo) {
+            $relImagenes[$relArticulo->id] = MprImagen::select('id', 'ruta', 'articulo')
+                ->where('articulo', $relArticulo->id)
+                ->where('estado', 1)
+                ->first();
         }
         $imagenes = MprImagen::where('articulo', $id)
             ->where('estado', 1)
             ->get();
 
         $comentarios = MadComentario::where('articulo', $id)
-        ->where('estado',1)
-        ->get();
+            ->where('estado', 1)
+            ->get();
 
-        return view ('mercancia.mprarticulos.show',compact('articulo','marca','tparticulos','comentarios','relacionados','relImagenes','imagenes'));
+        return view('mercancia.mprarticulos.show', compact('articulo', 'marca', 'tparticulos', 'comentarios', 'relacionados', 'relImagenes', 'imagenes'));
     }
 
     public function showAdmin($id)
     {
-        $articulos = MprArticulo::findOrFail($id);
-        $tmarticulos = MprMarca::all();
-        $tparticulos = MprtpArticulo::all();
-        $imagenes = MprImagen::where('articulo', $id)
-        ->where('estado', 1)
-        ->get();
-        $comentarios = MadComentario::where('articulo', $id)
-        ->where('estado',1)
-        ->get();
 
-        return view ('mercancia.mprarticulos.AdminShow',compact('articulos','tmarticulos','tparticulos','imagenes','comentarios'));
+        if (Auth::guard('usuarios')->user()) {
+            $ctru = new musUsuarioController();
+            if ($ctru->getPermisoInv()->mostrar == 1) {
+                $articulos = MprArticulo::findOrFail($id);
+                $tmarticulos = MprMarca::all();
+                $tparticulos = MprtpArticulo::all();
+                $imagenes = MprImagen::where('articulo', $id)
+                    ->where('estado', 1)
+                    ->get();
+                $comentarios = MadComentario::where('articulo', $id)
+                    ->where('estado', 1)
+                    ->get();
+
+                return view('mercancia.mprarticulos.AdminShow', compact('articulos', 'tmarticulos', 'tparticulos', 'imagenes', 'comentarios'));
+            }
+        }
+        return redirect()->route('homeAdmin');
     }
 
     /**
@@ -143,10 +179,17 @@ class MprArticuloController extends Controller
      */
     public function edit($id)
     {
-        $articulos = MprArticulo::findOrFail($id);
-        $tmarticulos = MprMarca::all();
-        $tparticulos = MprtpArticulo::all();
-        return view ('mercancia.mprarticulos.editar',compact('articulos','tmarticulos','tparticulos'));
+
+        if (Auth::guard('usuarios')->user()) {
+            $ctru = new musUsuarioController();
+            if ($ctru->getPermisoInv()->editar == 1) {
+                $articulos = MprArticulo::findOrFail($id);
+                $tmarticulos = MprMarca::all();
+                $tparticulos = MprtpArticulo::all();
+                return view('mercancia.mprarticulos.editar', compact('articulos', 'tmarticulos', 'tparticulos'));
+            }
+        }
+        return redirect()->route('homeAdmin');
     }
 
     /**
@@ -158,14 +201,21 @@ class MprArticuloController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $fechaActulizacion = new MprFechaUpdateContoller();
-        
-        $datos = $request->all();
-        $articulos = MprArticulo::findOrFail($id);
-        $articulos->update($datos);
-        $articulos->factualizado =$fechaActulizacion->fecha();
-        $articulos->save();
-        return redirect("adminVerArtidulo/$id");
+
+        if (Auth::guard('usuarios')->user()) {
+            $ctru = new musUsuarioController();
+            if ($ctru->getPermisoInv()->editar == 1) {
+                $fechaActulizacion = new MprFechaUpdateContoller();
+
+                $datos = $request->all();
+                $articulos = MprArticulo::findOrFail($id);
+                $articulos->update($datos);
+                $articulos->factualizado = $fechaActulizacion->fecha();
+                $articulos->save();
+                return redirect("adminVerArtidulo/$id");
+            }
+        }
+        return redirect()->route('homeAdmin');
     }
 
     /**
@@ -177,22 +227,27 @@ class MprArticuloController extends Controller
 
     public function destroy($id)
     {
-        $datos = MprArticulo::findOrFail($id);
-        $fechaActulizacion = new MprFechaUpdateContoller();
 
-        $imagenes = MprImagen::select('id','ruta','estado')->where('estado',1)->where('articulo',$id)->get();
+        if (Auth::guard('usuarios')->user()) {
+            $ctru = new musUsuarioController();
+            if ($ctru->getPermisoInv()->eliminar == 1) {
+                $datos = MprArticulo::findOrFail($id);
+                $fechaActulizacion = new MprFechaUpdateContoller();
 
-        foreach ($imagenes as $imagen){
-            $ctri = new MprImagenController();
-            $ctri->destroy($imagen->id);
+                $imagenes = MprImagen::select('id', 'ruta', 'estado')->where('estado', 1)->where('articulo', $id)->get();
+
+                foreach ($imagenes as $imagen) {
+                    $ctri = new MprImagenController();
+                    $ctri->destroy($imagen->id);
+                }
+
+                $datos['factualizado'] = $fechaActulizacion->fecha();
+                $datos['estado'] = 0;
+                $datos->save();
+
+                return redirect()->route('adminArticulos');
+            }
         }
-
-        $datos['factualizado'] = $fechaActulizacion->fecha();
-        $datos['estado']=0;
-        $datos->save();
-
-        return redirect()->route('adminArticulos');
+        return redirect()->route('homeAdmin');
     }
-
-
 }
